passform2_agent_backend/pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "passform_agent_backend"
version = "0.1.0"
description = "Backend agent for processing ROS2 data and providing an API interface"
readme = "README.md"
license = {text = "Apache-2.0"}
dependencies = [
    "setuptools",
    "rclpy",
    "fastapi",
    "uvicorn",
    "pydantic",
    "pydantic-settings",
    "ython-socketio"
]

[project.scripts]
backend = "passform_agent_backend.main:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["passform_agent_backend*"]

[tool.setuptools.data-files]
"share/ament_index/resource_index/packages" = ["resource/passform_agent_frontend"]
"share/passform_agent_backend" = ["package.xml"]
"share/passform_agent_backend/launch" = ["launch/backend_launch.py"]